TosaLayerwiseConstantFold: Merge all patterns into TosaFolders.cpp
