+++ mlir/lib/Dialect/Tosa/Transforms/CMakeLists.txt
@@ -2,13 +2,6 @@ add_mlir_dialect_library(MLIRTosaTransforms
   TosaDecomposeTransposeConv.cpp
   TosaDecomposeConv2D.cpp
   TosaDecomposeDepthwise.cpp
-  TosaFoldCommon.cpp
-  TosaFoldConstantAdd.cpp
-  TosaFoldConstantCast.cpp
-  TosaFoldConstantClamp.cpp
-  TosaFoldConstantMul.cpp
-  TosaFoldConstantPow.cpp
-  TosaFoldConstantRSQRT.cpp
   TosaFolders.cpp
   TosaInferShapes.cpp
   TosaLayerwiseConstantFoldPass.cpp
